Give the Windows build the same window, by cross-building Cairo rather than porting the drawing
The window is 161 Cairo calls and 29 X11 calls. Rewriting the drawing against a
Windows drawing API would have meant maintaining two of them forever and the
two would have drifted, so instead Cairo itself is cross-built with its win32
backend and only the windowing changes. The drawing code is untouched, and the
Windows window is the Linux window: same layout, same pixels.

cmake/build-windows-cairo.sh builds pixman and Cairo for MinGW with everything
else switched off -- no X11, no freetype or fontconfig since the win32 backend
brings its own font handling, no PNG, no zlib, no glib. That leaves pixman as
the only dependency and it has none of its own. The resulting .clap imports
nothing but stock Windows DLLs.

What actually differs between the platforms is now a short list: making a
window, parenting it, showing it, pumping events, and where the finished frame
is blitted. The event handlers were made platform-neutral first, so X11 and
Win32 both translate into the same onPointerDown, onMotion and onSaveKey, and
everything past that point is shared.

Keyboard focus is the one place Windows is simply better. On X11 the save field
has to take a keyboard grab because an embedded window is not given the focus
and CLAP cannot ask for it; on Windows a child window may take the focus, so it
does, the host keeps its shortcuts, and focus returns on its own.

Fixes a truncation bug found while wiring the handle through: the Gui interface
took a native window handle as `unsigned long`, which is 32 bits on 64-bit
Windows, so an HWND would have been quietly cut in half. It is `uintptr_t` now.

Verified on Linux at every step -- self-test 45 checks, validator 38 passed, 0
failed -- and under Wine for the Windows build, where the self-test also passes
all 45 against the DLL with the window compiled in.

+++ rainyday/src/gui/gui.h
@@ -60,8 +60,10 @@ public:
 
    // Creates the window, unmapped and unparented.
    virtual bool open() = 0;
-   virtual bool embed(unsigned long parentWindow) = 0;
-   virtual bool setTransientFor(unsigned long parentWindow) = 0;
+   // A native window handle. Not `unsigned long`: that is 32 bits on 64-bit
+   // Windows, which would quietly truncate an HWND.
+   virtual bool embed(uintptr_t parentWindow) = 0;
+   virtual bool setTransientFor(uintptr_t parentWindow) = 0;
    virtual void setTitle(const char *title) = 0;
    virtual void setScale(double scale) = 0;
    virtual void size(uint32_t *width, uint32_t *height) const = 0;
